+++ include/UDRefl/Util.h
@@ -491,6 +491,8 @@ namespace Ubpa::UDRefl {
 	concept container_front = requires(T t) { t.front(); };
 	template<typename T>
 	concept container_back = requires(T t) { t.back(); };
+	template<typename T>
+	concept container_top = requires(T t) { t.top(); };
 
 	// - capacity
 
@@ -607,6 +609,16 @@ namespace Ubpa::UDRefl {
 
 	template<typename T>
 	concept container_pop_back = requires(T t) { t.pop_back(); };
+
+	template<typename T, typename U>
+	concept container_push = requires(T t, U u) { t.push(std::forward<U>(u)); };
+	template<typename T>
+	concept container_push_clvalue = container_value_type<T> && container_push<T, const typename T::value_type&>;
+	template<typename T>
+	concept container_push_rvalue = container_value_type<T> && container_push<T, typename T::value_type&&>;
+
+	template<typename T>
+	concept container_pop = requires(T t) { t.pop(); };
 
 	template<typename T>
 	concept container_swap = requires(T lhs, T rhs) { std::swap(lhs, rhs); };